Use openpgp.js to encrypt with a hardcoded password
This is a very simple proof that basically openpgp.js runs at all.
It uses a hardcoded password to replace the string reversal
encryption and decryption functions.

At the time of this commit, openpgp was included using Bower.  I
don't like that, so rather than include it in this commit, I am
going to install it as a submodule next commit.