Encrypt/Decrypt using a PGP keypair
Use openpgp to generate a keypair for encryption and decryption. Once a
keypair is generated, stash it alongside the passwords in your crypt.
There is some cleanup to do with `browse` so that it doesn't list the cert
URL with the passwords.

While the keypair is protected by a passphrase, this is currently just
hardcoded. Immediate next steps would be to allow for input of a
passphrase and uploading of a personal cert file.

Also, I wasn't comfortable including the entire openpgpjs project in a
directory that gets served from a webserver.  Too many extra files lying
around to potentially cause trouble.  So, rather than including a
submodule, I ended up just including the four source files directly.