Refactor into a vanilla MVC style
This refactor probably doesn't do much in the long run, but it certainly
makes me feel really good. :-)

In any case, I restructured the crypt into a model and the console actions
into a simple view, glued to each other using the main file and promise
resolutions. This has the immediately effect of hiding the passphrase from
the view, for example.

The pattern is modeled after my tadhack-bitsip project, which is in turn
modeled after the Vanilla TodoJS demo (which looks like has since been
made much less Vanilla, IMHO).