Browse and display stored names on load!
I continued to add more and more wiring a la tadhack-bitsip. Replacing the
PasswordConsoleView with a view that actually ties to the UI, this so far
loads in stored password names and displays them in the list. How awesome
is that!

The main pieces of this are a simple template function, which can generate
new DOM nodes based on existing (hidden) ones, plus wiring in all the
right places.  The main.js file is what actually calls browse, once
everything is hooked up.