Cache static files so the app works offline
Static files are cached using the serviceWorker, so that as long as you
get an initial app load, you can continue to use it offline. Some files
are fetched up front, which I don't think works right now (based on path
names), but that can be cleaned up once it is built to a permanent
location.

Caching is still treated as network-first, cache-second, so that I don't
get overly annoyed as I continue to update various files.

I have to say, this is pretty awesome!