Beginning work on WebRTC DataChannel sync
In order to sync passwords between devices, we need to establish a
signaling connection and use that to then set up a DataChannel. I reached
for the arguably heavy but definitely familiar SIP.js for this purpose.

This commit lays the groundwork, with a new SyncModel that can connect
data channels between two endpoints.  There are still major security holes
to plug, most likely involving a simple manual verification of
fingerprints.

This is purely console-based at the moment.  Still unsure how the actual
syncing UI works, especially for two non-empty password libraries.