Sync is online!
Password sync is now pretty much fully functional! You can connect, verify
matching fingerprints, and transfer both your passwords as well as your
cert.

Be aware!  Syncing your certificate complete blows away whatever was on
the other end, so any passwords that were there using a different cert
will be broken.  This is intended to work between crypts with the same
cert file.

All passwords and certs are transfered encrypted, and even the plaintext
data (like password names or the type of message) are in a verified
peer-to-peer connection.  Security!